Cleanup landing app * Load districts on mount * Move theming to App.js
