+++ src/app/App.js
@@ -1,22 +1,25 @@
-import {h} from 'preact';
-
 import {
   HashRouter,
-  Route,
+  Link,
   PropsRoute,
-  Link, Redirect, Switch
+  Redirect,
+  Route,
+  Switch
 } from 'react-router-dom';
-import BillTrackerContainer from 'components/bills/tracker/BillTrackerContainer';
 
 import BillFilterContainer from 'components/bills/filter/BillFilterContainer';
 import BillDetailContainer from 'components/bills/detail/BillDetailContainer';
 import BillSearchContainer from 'components/bills/search/BillSearchContainer';
-import BillSearchContainerDynamic from 'components/bills/search/BillSearchContainerDynamic';
 
 import LandingApp from 'components/landing_app';
-import { ROOT_PATH } from 'local_constants';
 
 import RaisedButton from 'material-ui/RaisedButton';
+
+import {withStyles} from 'material-ui/styles';
+import lightBaseTheme from 'material-ui/styles/baseThemes/lightBaseTheme';
+import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
+import getMuiTheme from 'material-ui/styles/getMuiTheme';
+import 'styles/style.less';
 
 const style = {
   margin: 12,
@@ -30,21 +33,21 @@ const Header = () => (
   </div>
 );
 
-const App = function (districtList) {
-  return  (
-  <HashRouter>
-    <div>
-      <Header />
-      <Switch>
-        <Route path="/find_rep" render={(props) => (
-          <LandingApp {...props} stateDistricts={districtList} /> )} />
-        <Route path="/bill_search" component={BillSearchContainer}/>
-        <Route path="/bill_filter" component={BillFilterContainer}/>
-        <Route path="/bills/:bill_name" component={BillDetailContainer}/>
-        <Redirect from="/" to="find_rep"/>
-      </Switch>
-    </div>
-  </HashRouter>
-)};
+const App = () => (
+  <MuiThemeProvider muiTheme={getMuiTheme(lightBaseTheme)}>
+    <HashRouter>
+      <div>
+        <Header />
+        <Switch>
+          <Route exact path="/find_rep" component={LandingApp} />
+          <Route path="/bill_search" component={BillSearchContainer}/>
+          <Route path="/bill_filter" component={BillFilterContainer}/>
+          <Route path="/bills/:bill_name" component={BillDetailContainer}/>
+          <Redirect from="/" to="find_rep"/>
+        </Switch>
+      </div>
+    </HashRouter>
+  </MuiThemeProvider>
+);
 
 export default App;
